Do some code cleanup
